@@ -1,11 +1,11 @@
 package creditscoreservice.it;
 
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
-import static org.junit.Assert.assertThat;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import javax.ws.rs.core.Response;
-import org.junit.Test;
+import jakarta.ws.rs.core.Response;
+import org.junit.jupiter.api.Test;
 
 public class CreditScoreIT extends IntegrationTestBase {
 
