@@ -1,7 +1,7 @@
 package specialmembershipservice.port.incoming.adapter.resources;
 
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotNull;
 
 public class SpecialMembershipDto {
 
