@@ -2,56 +2,57 @@ package welcomememberemailservice.it;
 
 import static io.dropwizard.testing.ResourceHelpers.resourceFilePath;
 import static org.awaitility.Awaitility.await;
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasSize;
-import static org.junit.Assert.assertThat;
 
-import com.github.charithe.kafka.EphemeralKafkaBroker;
-import com.github.charithe.kafka.KafkaJunitRule;
-import io.dropwizard.testing.junit.DropwizardAppRule;
+import io.dropwizard.testing.junit5.DropwizardAppExtension;
 import java.lang.invoke.MethodHandles;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
-import org.junit.Before;
-import org.junit.ClassRule;
-import org.junit.rules.RuleChain;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Order;
+import org.junit.jupiter.api.extension.RegisterExtension;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.subethamail.wiser.WiserMessage;
 import welcomememberemailservice.bootstrap.WelcomeMemberEmailServiceApplication;
 import welcomememberemailservice.bootstrap.WelcomeMemberEmailServiceConfiguration;
 import welcomememberemailservice.it.kafka.KafkaConsumerOffsets;
-import welcomememberemailservice.it.smtp.SmtpServerRule;
+import welcomememberemailservice.it.kafka.KafkaContainerExtension;
+import welcomememberemailservice.it.smtp.SmtpServerExtension;
 
 public abstract class IntegrationTestBase {
 
   private static final Logger LOG = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
   private static final String INTEGRATION_YML = resourceFilePath("integration.yml");
   private static final String KAFKA_HOST = "localhost";
-  private static final int KAFKA_PORT = 9092;
+  private static final int KAFKA_PORT = 19092;
   private static final String SPECIAL_MEMBERSHIP_TOPIC = "special-membership-topic";
   private static final String WELCOME_EMAIL_GROUP_ID = "welcome-member-email-consumer";
   private static final int SMTP_SERVER_PORT = 2525;
 
-  private static final EphemeralKafkaBroker KAFKA_BROKER = EphemeralKafkaBroker.create(KAFKA_PORT);
-  private static final KafkaJunitRule KAFKA_RULE = new KafkaJunitRule(KAFKA_BROKER)
-      .waitForStartup();
-  private static final SmtpServerRule SMTP_SERVER_RULE = new SmtpServerRule(SMTP_SERVER_PORT);
-  private static final DropwizardAppRule<WelcomeMemberEmailServiceConfiguration> SERVICE_RULE =
-      new DropwizardAppRule<>(WelcomeMemberEmailServiceApplication.class, INTEGRATION_YML);
+  @RegisterExtension
+  @Order(0)
+  private static final SmtpServerExtension SMTP_SERVER_RULE = new SmtpServerExtension(
+      SMTP_SERVER_PORT);
 
-  @ClassRule
-  public static final RuleChain RULES = RuleChain
-      .outerRule(KAFKA_RULE)
-      .around(SMTP_SERVER_RULE)
-      .around(SERVICE_RULE);
+  @RegisterExtension
+  @Order(1)
+  private static final KafkaContainerExtension KAFKA_RULE = new KafkaContainerExtension(
+      KAFKA_HOST, KAFKA_PORT);
+
+  @RegisterExtension
+  @Order(Integer.MAX_VALUE)
+  private static final DropwizardAppExtension<WelcomeMemberEmailServiceConfiguration> SERVICE_RULE =
+      new DropwizardAppExtension<>(WelcomeMemberEmailServiceApplication.class, INTEGRATION_YML);
 
   private List<WiserMessage> emailsBeforeTest;
 
-  @Before
-  public void setUp() throws Exception {
-    emailsBeforeTest = getEmails();
+  @BeforeEach
+  public void setUp() {
+    this.emailsBeforeTest = getEmails();
   }
 
   private List<WiserMessage> getEmails() {
@@ -77,12 +78,12 @@ public abstract class IntegrationTestBase {
 
   protected void publishMessageAndWaitToBeConsumed(String topic, String message, String groupId) {
     KafkaConsumerOffsets kafkaConsumerOffsets =
-        new KafkaConsumerOffsets(KAFKA_HOST, KAFKA_RULE.helper().kafkaPort(), groupId);
+        new KafkaConsumerOffsets(KAFKA_HOST, KAFKA_RULE.getPort(), groupId);
 
     final long previousOffset = Math.max(kafkaConsumerOffsets.readOffset(topic), 0);
 
     LOG.info("Publishing message {} to topic {}", message, topic);
-    KAFKA_RULE.helper().produceStrings(topic, message);
+    KAFKA_RULE.produceStrings(topic, message);
 
     LOG.info("Waiting for message to be consumed from topic {}", topic);
     await().atMost(Duration.ofSeconds(5))
