@@ -1,26 +1,22 @@
 package welcomememberemailservice.it.pacts;
 
-import static java.nio.charset.StandardCharsets.UTF_8;
 import static welcomememberemailservice.it.pacts.PactConstants.SPECIAL_MEMBERSHIP_SERVICE;
 import static welcomememberemailservice.it.pacts.PactConstants.WELCOME_MEMBER_EMAIL_SERVICE;
 
 import au.com.dius.pact.consumer.MessagePactBuilder;
 import au.com.dius.pact.consumer.dsl.PactDslJsonBody;
-import au.com.dius.pact.consumer.junit.MessagePactProviderRule;
-import au.com.dius.pact.consumer.junit.PactVerification;
+import au.com.dius.pact.consumer.junit5.PactTestFor;
+import au.com.dius.pact.consumer.junit5.ProviderType;
+import au.com.dius.pact.core.model.PactSpecVersion;
+import au.com.dius.pact.core.model.V4Interaction;
 import au.com.dius.pact.core.model.annotations.Pact;
 import au.com.dius.pact.core.model.messaging.MessagePact;
 import java.util.HashMap;
 import java.util.Map;
-import org.junit.Rule;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import welcomememberemailservice.it.IntegrationTestBase;
 
 public class MemberSignedUpEventPactContracts extends IntegrationTestBase {
-
-  @Rule
-  public final MessagePactProviderRule specialMembershipServiceRule =
-      new MessagePactProviderRule(SPECIAL_MEMBERSHIP_SERVICE, this);
 
   @Pact(consumer = WELCOME_MEMBER_EMAIL_SERVICE, provider = SPECIAL_MEMBERSHIP_SERVICE)
   public MessagePact newMemberTonyStark(MessagePactBuilder builder) {
@@ -37,10 +33,11 @@ public class MemberSignedUpEventPactContracts extends IntegrationTestBase {
         .toPact();
   }
 
-  @PactVerification(value = SPECIAL_MEMBERSHIP_SERVICE, fragment = "newMemberTonyStark")
   @Test
-  public void sendWelcomeEmailToTonyStark() throws Exception {
-    String memberSignedUpEvent = new String(specialMembershipServiceRule.getMessage(), UTF_8);
+  @PactTestFor(providerName = SPECIAL_MEMBERSHIP_SERVICE, pactMethod = "newMemberTonyStark",
+      providerType = ProviderType.ASYNCH, pactVersion = PactSpecVersion.V4)
+  public void sendWelcomeEmailToTonyStark(V4Interaction.AsynchronousMessage message) {
+    String memberSignedUpEvent = message.contentsAsString();
     publishMembershipMessageAndWaitToBeConsumed(memberSignedUpEvent);
     assertAnEmailWasSent();
   }
