@@ -4,19 +4,15 @@ import static creditscoreservice.it.pacts.PactConstants.CREDIT_SCORE_SERVICE;
 import static creditscoreservice.it.pacts.PactConstants.PACT_BROKER_PORT;
 import static creditscoreservice.it.pacts.PactConstants.PACT_BROKER_URL;
 
-import au.com.dius.pact.provider.junit.IgnoreNoPactsToVerify;
-import au.com.dius.pact.provider.junit.Provider;
-import au.com.dius.pact.provider.junit.RestPactRunner;
-import au.com.dius.pact.provider.junit.State;
-import au.com.dius.pact.provider.junit.loader.PactBroker;
-import au.com.dius.pact.provider.junit.loader.PactBrokerAuth;
-import au.com.dius.pact.provider.junit.target.HttpTarget;
-import au.com.dius.pact.provider.junit.target.Target;
-import au.com.dius.pact.provider.junit.target.TestTarget;
+import au.com.dius.pact.provider.junit5.HttpTestTarget;
+import au.com.dius.pact.provider.junitsupport.IgnoreNoPactsToVerify;
+import au.com.dius.pact.provider.junitsupport.Provider;
+import au.com.dius.pact.provider.junitsupport.State;
+import au.com.dius.pact.provider.junitsupport.loader.PactBroker;
+import au.com.dius.pact.provider.junitsupport.loader.PactBrokerAuth;
+import au.com.dius.pact.provider.junitsupport.target.TestTarget;
 import creditscoreservice.it.IntegrationTestBase;
-import org.junit.runner.RunWith;
 
-@RunWith(RestPactRunner.class)
 @Provider(CREDIT_SCORE_SERVICE)
 @IgnoreNoPactsToVerify
 @PactBroker(
@@ -29,7 +25,7 @@ import org.junit.runner.RunWith;
 public class HttpPactVerifications extends IntegrationTestBase {
 
   @TestTarget
-  public final Target target = new HttpTarget(SERVICE_RULE.getLocalPort());
+  public final HttpTestTarget target = new HttpTestTarget("localhost", SERVICE_RULE.getLocalPort());
 
   @State("There is a tony.stark@example.com")
   public void tonyStarkCreditScore() {
